fix: handle empty dashboard eb calculation
(cherry picked from commit 6ba80671c14b20ba40f1f68cf0ecc1626cc3efef)

+++ backend/pkg/exporter/modules/dashboard_data.go
@@ -881,7 +881,7 @@ func (d *dashboardData) fetchEpochValidatorStates(epochStart uint64, epochEnd ui
 			slots = append(slots, 0)
 			continue
 		}
-		slots = append(slots, uint64(i)*utils.Config.Chain.ClConfig.SlotsPerEpoch-1)
+		slots = append(slots, i*utils.Config.Chain.ClConfig.SlotsPerEpoch-1)
 	}
 	writeMutex := &sync.Mutex{}
 	d.log.Debugf("fetching states for epochs %d to %d using slots %v", epochStart, epochEnd, slots)
